% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{AutoForwarding}
\alias{AutoForwarding}
\title{Gmail API Objects 
Access Gmail mailboxes including sending user email.}
\usage{
AutoForwarding(disposition = NULL, emailAddress = NULL, enabled = NULL)
}
\arguments{
\item{disposition}{The state that a message should be left in after it has been forwarded}

\item{emailAddress}{Email address to which all incoming messages are forwarded}

\item{enabled}{Whether all incoming mail is automatically forwarded to another address}
}
\value{
AutoForwarding object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-04 00:01:53
filename: /Users/mark/dev/R/autoGoogleAPI/googlegmailv1.auto/R/gmail_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AutoForwarding Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Auto-forwarding settings for an account.
}

