% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{BatchDeleteMessagesRequest}
\alias{BatchDeleteMessagesRequest}
\title{BatchDeleteMessagesRequest Object}
\usage{
BatchDeleteMessagesRequest(ids = NULL)
}
\arguments{
\item{ids}{The IDs of the messages to delete}
}
\value{
BatchDeleteMessagesRequest object
}
\description{
BatchDeleteMessagesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

