% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{BatchModifyMessagesRequest}
\alias{BatchModifyMessagesRequest}
\title{BatchModifyMessagesRequest Object}
\usage{
BatchModifyMessagesRequest(addLabelIds = NULL, ids = NULL,
  removeLabelIds = NULL)
}
\arguments{
\item{addLabelIds}{A list of label IDs to add to messages}

\item{ids}{The IDs of the messages to modify}

\item{removeLabelIds}{A list of label IDs to remove from messages}
}
\value{
BatchModifyMessagesRequest object
}
\description{
BatchModifyMessagesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
