% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{Draft}
\alias{Draft}
\title{Draft Object}
\usage{
Draft(id = NULL, message = NULL)
}
\arguments{
\item{id}{The immutable ID of the draft}

\item{message}{The message content of the draft}
}
\value{
Draft object
}
\description{
Draft Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A draft email in the user's mailbox.
}

