% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{Filter}
\alias{Filter}
\title{Filter Object}
\usage{
Filter(action = NULL, criteria = NULL, id = NULL)
}
\arguments{
\item{action}{Action that the filter performs}

\item{criteria}{Matching criteria for the filter}

\item{id}{The server assigned ID of the filter}
}
\value{
Filter object
}
\description{
Filter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Resource definition for Gmail filters. Filters apply to specific messages instead of an entire email thread.
}
