% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{FilterAction}
\alias{FilterAction}
\title{FilterAction Object}
\usage{
FilterAction(addLabelIds = NULL, forward = NULL, removeLabelIds = NULL)
}
\arguments{
\item{addLabelIds}{List of labels to add to the message}

\item{forward}{Email address that the message should be forwarded to}

\item{removeLabelIds}{List of labels to remove from the message}
}
\value{
FilterAction object
}
\description{
FilterAction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of actions to perform on a message.
}
