% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{FilterCriteria}
\alias{FilterCriteria}
\title{FilterCriteria Object}
\usage{
FilterCriteria(excludeChats = NULL, from = NULL, hasAttachment = NULL,
  negatedQuery = NULL, query = NULL, size = NULL, sizeComparison = NULL,
  subject = NULL, to = NULL)
}
\arguments{
\item{excludeChats}{Whether the response should exclude chats}

\item{from}{The sender's display name or email address}

\item{hasAttachment}{Whether the message has any attachment}

\item{negatedQuery}{Only return messages not matching the specified query}

\item{query}{Only return messages matching the specified query}

\item{size}{The size of the entire RFC822 message in bytes, including all headers and attachments}

\item{sizeComparison}{How the message size in bytes should be in relation to the size field}

\item{subject}{Case-insensitive phrase found in the message's subject}

\item{to}{The recipient's display name or email address}
}
\value{
FilterCriteria object
}
\description{
FilterCriteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message matching criteria.
}

