% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ForwardingAddress}
\alias{ForwardingAddress}
\title{ForwardingAddress Object}
\usage{
ForwardingAddress(forwardingEmail = NULL, verificationStatus = NULL)
}
\arguments{
\item{forwardingEmail}{An email address to which messages can be forwarded}

\item{verificationStatus}{Indicates whether this address has been verified and is usable for forwarding}
}
\value{
ForwardingAddress object
}
\description{
ForwardingAddress Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Settings for a forwarding address.
}
