% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_functions.R
\docType{package}
\name{gmail_googleAuthR}
\alias{gmail_googleAuthR}
\alias{gmail_googleAuthR-package}
\title{Gmail API
Access Gmail mailboxes including sending user email.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-04 00:01:53
filename: /Users/mark/dev/R/autoGoogleAPI/googlegmailv1.auto/R/gmail_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://mail.google.com/
\item https://www.googleapis.com/auth/gmail.compose
\item https://www.googleapis.com/auth/gmail.insert
\item https://www.googleapis.com/auth/gmail.labels
\item https://www.googleapis.com/auth/gmail.modify
\item https://www.googleapis.com/auth/gmail.readonly
\item https://www.googleapis.com/auth/gmail.send
\item https://www.googleapis.com/auth/gmail.settings.basic
\item https://www.googleapis.com/auth/gmail.settings.sharing
}
}

