% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{History}
\alias{History}
\title{History Object}
\usage{
History(id = NULL, labelsAdded = NULL, labelsRemoved = NULL,
  messages = NULL, messagesAdded = NULL, messagesDeleted = NULL)
}
\arguments{
\item{id}{The mailbox sequence ID}

\item{labelsAdded}{Labels added to messages in this history record}

\item{labelsRemoved}{Labels removed from messages in this history record}

\item{messages}{List of messages changed in this history record}

\item{messagesAdded}{Messages added to the mailbox in this history record}

\item{messagesDeleted}{Messages deleted (not Trashed) from the mailbox in this history record}
}
\value{
History object
}
\description{
History Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A record of a change to the user's mailbox. Each history change may affect multiple messages in multiple ways.
}

