% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{HistoryLabelAdded}
\alias{HistoryLabelAdded}
\title{HistoryLabelAdded Object}
\usage{
HistoryLabelAdded(labelIds = NULL, message = NULL)
}
\arguments{
\item{labelIds}{Label IDs added to the message}

\item{message}{No description}
}
\value{
HistoryLabelAdded object
}
\description{
HistoryLabelAdded Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
