% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{HistoryLabelRemoved}
\alias{HistoryLabelRemoved}
\title{HistoryLabelRemoved Object}
\usage{
HistoryLabelRemoved(labelIds = NULL, message = NULL)
}
\arguments{
\item{labelIds}{Label IDs removed from the message}

\item{message}{No description}
}
\value{
HistoryLabelRemoved object
}
\description{
HistoryLabelRemoved Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
