% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{HistoryMessageAdded}
\alias{HistoryMessageAdded}
\title{HistoryMessageAdded Object}
\usage{
HistoryMessageAdded(message = NULL)
}
\arguments{
\item{message}{No description}
}
\value{
HistoryMessageAdded object
}
\description{
HistoryMessageAdded Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

