% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{HistoryMessageDeleted}
\alias{HistoryMessageDeleted}
\title{HistoryMessageDeleted Object}
\usage{
HistoryMessageDeleted(message = NULL)
}
\arguments{
\item{message}{No description}
}
\value{
HistoryMessageDeleted object
}
\description{
HistoryMessageDeleted Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
