% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ImapSettings}
\alias{ImapSettings}
\title{ImapSettings Object}
\usage{
ImapSettings(autoExpunge = NULL, enabled = NULL, expungeBehavior = NULL,
  maxFolderSize = NULL)
}
\arguments{
\item{autoExpunge}{If this value is true, Gmail will immediately expunge a message when it is marked as deleted in IMAP}

\item{enabled}{Whether IMAP is enabled for the account}

\item{expungeBehavior}{The action that will be executed on a message when it is marked as deleted and expunged from the last visible IMAP folder}

\item{maxFolderSize}{An optional limit on the number of messages that an IMAP folder may contain}
}
\value{
ImapSettings object
}
\description{
ImapSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
IMAP settings for an account.
}
