% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{Label}
\alias{Label}
\title{Label Object}
\usage{
Label(id = NULL, labelListVisibility = NULL, messageListVisibility = NULL,
  messagesTotal = NULL, messagesUnread = NULL, name = NULL,
  threadsTotal = NULL, threadsUnread = NULL, type = NULL)
}
\arguments{
\item{id}{The immutable ID of the label}

\item{labelListVisibility}{The visibility of the label in the label list in the Gmail web interface}

\item{messageListVisibility}{The visibility of the label in the message list in the Gmail web interface}

\item{messagesTotal}{The total number of messages with the label}

\item{messagesUnread}{The number of unread messages with the label}

\item{name}{The display name of the label}

\item{threadsTotal}{The total number of threads with the label}

\item{threadsUnread}{The number of unread threads with the label}

\item{type}{The owner type for the label}
}
\value{
Label object
}
\description{
Label Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels are used to categorize messages and threads within the user's mailbox.
}

