% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ListDraftsResponse}
\alias{ListDraftsResponse}
\title{ListDraftsResponse Object}
\usage{
ListDraftsResponse(drafts = NULL, nextPageToken = NULL,
  resultSizeEstimate = NULL)
}
\arguments{
\item{drafts}{List of drafts}

\item{nextPageToken}{Token to retrieve the next page of results in the list}

\item{resultSizeEstimate}{Estimated total number of results}
}
\value{
ListDraftsResponse object
}
\description{
ListDraftsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

