% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ListFiltersResponse}
\alias{ListFiltersResponse}
\title{ListFiltersResponse Object}
\usage{
ListFiltersResponse(filter = NULL)
}
\arguments{
\item{filter}{List of a user's filters}
}
\value{
ListFiltersResponse object
}
\description{
ListFiltersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the ListFilters method.
}

