% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ListForwardingAddressesResponse}
\alias{ListForwardingAddressesResponse}
\title{ListForwardingAddressesResponse Object}
\usage{
ListForwardingAddressesResponse(forwardingAddresses = NULL)
}
\arguments{
\item{forwardingAddresses}{List of addresses that may be used for forwarding}
}
\value{
ListForwardingAddressesResponse object
}
\description{
ListForwardingAddressesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the ListForwardingAddresses method.
}
