% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ListHistoryResponse}
\alias{ListHistoryResponse}
\title{ListHistoryResponse Object}
\usage{
ListHistoryResponse(history = NULL, historyId = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{history}{List of history records}

\item{historyId}{The ID of the mailbox's current history record}

\item{nextPageToken}{Page token to retrieve the next page of results in the list}
}
\value{
ListHistoryResponse object
}
\description{
ListHistoryResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

