% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ListLabelsResponse}
\alias{ListLabelsResponse}
\title{ListLabelsResponse Object}
\usage{
ListLabelsResponse(labels = NULL)
}
\arguments{
\item{labels}{List of labels}
}
\value{
ListLabelsResponse object
}
\description{
ListLabelsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

