% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ListMessagesResponse}
\alias{ListMessagesResponse}
\title{ListMessagesResponse Object}
\usage{
ListMessagesResponse(messages = NULL, nextPageToken = NULL,
  resultSizeEstimate = NULL)
}
\arguments{
\item{messages}{List of messages}

\item{nextPageToken}{Token to retrieve the next page of results in the list}

\item{resultSizeEstimate}{Estimated total number of results}
}
\value{
ListMessagesResponse object
}
\description{
ListMessagesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
