% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ListSendAsResponse}
\alias{ListSendAsResponse}
\title{ListSendAsResponse Object}
\usage{
ListSendAsResponse(sendAs = NULL)
}
\arguments{
\item{sendAs}{List of send-as aliases}
}
\value{
ListSendAsResponse object
}
\description{
ListSendAsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the ListSendAs method.
}

