% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ListSmimeInfoResponse}
\alias{ListSmimeInfoResponse}
\title{ListSmimeInfoResponse Object}
\usage{
ListSmimeInfoResponse(smimeInfo = NULL)
}
\arguments{
\item{smimeInfo}{List of SmimeInfo}
}
\value{
ListSmimeInfoResponse object
}
\description{
ListSmimeInfoResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
