% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ListThreadsResponse}
\alias{ListThreadsResponse}
\title{ListThreadsResponse Object}
\usage{
ListThreadsResponse(nextPageToken = NULL, resultSizeEstimate = NULL,
  threads = NULL)
}
\arguments{
\item{nextPageToken}{Page token to retrieve the next page of results in the list}

\item{resultSizeEstimate}{Estimated total number of results}

\item{threads}{List of threads}
}
\value{
ListThreadsResponse object
}
\description{
ListThreadsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

