% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{Message}
\alias{Message}
\title{Message Object}
\usage{
Message(historyId = NULL, id = NULL, internalDate = NULL,
  labelIds = NULL, payload = NULL, raw = NULL, sizeEstimate = NULL,
  snippet = NULL, threadId = NULL)
}
\arguments{
\item{historyId}{The ID of the last history record that modified this message}

\item{id}{The immutable ID of the message}

\item{internalDate}{The internal message creation timestamp (epoch ms), which determines ordering in the inbox}

\item{labelIds}{List of IDs of labels applied to this message}

\item{payload}{The parsed email structure in the message parts}

\item{raw}{The entire email message in an RFC 2822 formatted and base64url encoded string}

\item{sizeEstimate}{Estimated size in bytes of the message}

\item{snippet}{A short part of the message text}

\item{threadId}{The ID of the thread the message belongs to}
}
\value{
Message object
}
\description{
Message Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An email message.
}

