% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{MessagePart}
\alias{MessagePart}
\title{MessagePart Object}
\usage{
MessagePart(body = NULL, filename = NULL, headers = NULL,
  mimeType = NULL, partId = NULL, parts = NULL)
}
\arguments{
\item{body}{The message part body for this part, which may be empty for container MIME message parts}

\item{filename}{The filename of the attachment}

\item{headers}{List of headers on this message part}

\item{mimeType}{The MIME type of the message part}

\item{partId}{The immutable ID of the message part}

\item{parts}{The child MIME message parts of this part}
}
\value{
MessagePart object
}
\description{
MessagePart Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single MIME message part.
}
