% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{MessagePartBody}
\alias{MessagePartBody}
\title{MessagePartBody Object}
\usage{
MessagePartBody(attachmentId = NULL, data = NULL, size = NULL)
}
\arguments{
\item{attachmentId}{When present, contains the ID of an external attachment that can be retrieved in a separate messages}

\item{data}{The body data of a MIME message part}

\item{size}{Total number of bytes in the body of the message part}
}
\value{
MessagePartBody object
}
\description{
MessagePartBody Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The body of a single MIME message part.
}

