% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{MessagePartHeader}
\alias{MessagePartHeader}
\title{MessagePartHeader Object}
\usage{
MessagePartHeader(name = NULL, value = NULL)
}
\arguments{
\item{name}{The name of the header before the : separator}

\item{value}{The value of the header after the : separator}
}
\value{
MessagePartHeader object
}
\description{
MessagePartHeader Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
