% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ModifyMessageRequest}
\alias{ModifyMessageRequest}
\title{ModifyMessageRequest Object}
\usage{
ModifyMessageRequest(addLabelIds = NULL, removeLabelIds = NULL)
}
\arguments{
\item{addLabelIds}{A list of IDs of labels to add to this message}

\item{removeLabelIds}{A list IDs of labels to remove from this message}
}
\value{
ModifyMessageRequest object
}
\description{
ModifyMessageRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
