% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{ModifyThreadRequest}
\alias{ModifyThreadRequest}
\title{ModifyThreadRequest Object}
\usage{
ModifyThreadRequest(addLabelIds = NULL, removeLabelIds = NULL)
}
\arguments{
\item{addLabelIds}{A list of IDs of labels to add to this thread}

\item{removeLabelIds}{A list of IDs of labels to remove from this thread}
}
\value{
ModifyThreadRequest object
}
\description{
ModifyThreadRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

