% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{PopSettings}
\alias{PopSettings}
\title{PopSettings Object}
\usage{
PopSettings(accessWindow = NULL, disposition = NULL)
}
\arguments{
\item{accessWindow}{The range of messages which are accessible via POP}

\item{disposition}{The action that will be executed on a message after it has been fetched via POP}
}
\value{
PopSettings object
}
\description{
PopSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
POP settings for an account.
}
