% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{Profile}
\alias{Profile}
\title{Profile Object}
\usage{
Profile(emailAddress = NULL, historyId = NULL, messagesTotal = NULL,
  threadsTotal = NULL)
}
\arguments{
\item{emailAddress}{The user's email address}

\item{historyId}{The ID of the mailbox's current history record}

\item{messagesTotal}{The total number of messages in the mailbox}

\item{threadsTotal}{The total number of threads in the mailbox}
}
\value{
Profile object
}
\description{
Profile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Profile for a Gmail user.
}
