% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{SendAs}
\alias{SendAs}
\title{SendAs Object}
\usage{
SendAs(displayName = NULL, isDefault = NULL, isPrimary = NULL,
  replyToAddress = NULL, sendAsEmail = NULL, signature = NULL,
  smtpMsa = NULL, treatAsAlias = NULL, verificationStatus = NULL)
}
\arguments{
\item{displayName}{A name that appears in the 'From:' header for mail sent using this alias}

\item{isDefault}{Whether this address is selected as the default 'From:' address in situations such as composing a new message or sending a vacation auto-reply}

\item{isPrimary}{Whether this address is the primary address used to login to the account}

\item{replyToAddress}{An optional email address that is included in a 'Reply-To:' header for mail sent using this alias}

\item{sendAsEmail}{The email address that appears in the 'From:' header for mail sent using this alias}

\item{signature}{An optional HTML signature that is included in messages composed with this alias in the Gmail web UI}

\item{smtpMsa}{An optional SMTP service that will be used as an outbound relay for mail sent using this alias}

\item{treatAsAlias}{Whether Gmail should  treat this address as an alias for the user's primary email address}

\item{verificationStatus}{Indicates whether this address has been verified for use as a send-as alias}
}
\value{
SendAs object
}
\description{
SendAs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Settings associated with a send-as alias, which can be either the primary login address associated with the account or a custom 'from' address. Send-as aliases correspond to the 'Send Mail As' feature in the web interface.
}
