% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{SmimeInfo}
\alias{SmimeInfo}
\title{SmimeInfo Object}
\usage{
SmimeInfo(encryptedKeyPassword = NULL, expiration = NULL, id = NULL,
  isDefault = NULL, issuerCn = NULL, pem = NULL, pkcs12 = NULL)
}
\arguments{
\item{encryptedKeyPassword}{Encrypted key password, when key is encrypted}

\item{expiration}{When the certificate expires (in milliseconds since epoch)}

\item{id}{The immutable ID for the SmimeInfo}

\item{isDefault}{Whether this SmimeInfo is the default one for this user's send-as address}

\item{issuerCn}{The S/MIME certificate issuer's common name}

\item{pem}{PEM formatted X509 concatenated certificate string (standard base64 encoding)}

\item{pkcs12}{PKCS#12 format containing a single private/public key pair and certificate chain}
}
\value{
SmimeInfo object
}
\description{
SmimeInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An S/MIME email config.
}
