% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{SmtpMsa}
\alias{SmtpMsa}
\title{SmtpMsa Object}
\usage{
SmtpMsa(host = NULL, password = NULL, port = NULL, securityMode = NULL,
  username = NULL)
}
\arguments{
\item{host}{The hostname of the SMTP service}

\item{password}{The password that will be used for authentication with the SMTP service}

\item{port}{The port of the SMTP service}

\item{securityMode}{The protocol that will be used to secure communication with the SMTP service}

\item{username}{The username that will be used for authentication with the SMTP service}
}
\value{
SmtpMsa object
}
\description{
SmtpMsa Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for communication with an SMTP service.
}
