% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{Thread}
\alias{Thread}
\title{Thread Object}
\usage{
Thread(historyId = NULL, id = NULL, messages = NULL, snippet = NULL)
}
\arguments{
\item{historyId}{The ID of the last history record that modified this thread}

\item{id}{The unique ID of the thread}

\item{messages}{The list of messages in the thread}

\item{snippet}{A short part of the message text}
}
\value{
Thread object
}
\description{
Thread Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A collection of messages representing a conversation.
}

