% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_functions.R
\name{users.watch}
\alias{users.watch}
\title{Set up or update a push notification watch on the given user mailbox.}
\usage{
users.watch(WatchRequest, userId)
}
\arguments{
\item{WatchRequest}{The \link{WatchRequest} object to pass to this method}

\item{userId}{The user's email address}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://mail.google.com/
\item https://www.googleapis.com/auth/gmail.modify
\item https://www.googleapis.com/auth/gmail.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://mail.google.com/, https://www.googleapis.com/auth/gmail.modify, https://www.googleapis.com/auth/gmail.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/gmail/api/}{Google Documentation}

Other WatchRequest functions: \code{\link{WatchRequest}}
}

