% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{VacationSettings}
\alias{VacationSettings}
\title{VacationSettings Object}
\usage{
VacationSettings(enableAutoReply = NULL, endTime = NULL,
  responseBodyHtml = NULL, responseBodyPlainText = NULL,
  responseSubject = NULL, restrictToContacts = NULL,
  restrictToDomain = NULL, startTime = NULL)
}
\arguments{
\item{enableAutoReply}{Flag that controls whether Gmail automatically replies to messages}

\item{endTime}{An optional end time for sending auto-replies (epoch ms)}

\item{responseBodyHtml}{Response body in HTML format}

\item{responseBodyPlainText}{Response body in plain text format}

\item{responseSubject}{Optional text to prepend to the subject line in vacation responses}

\item{restrictToContacts}{Flag that determines whether responses are sent to recipients who are not in the user's list of contacts}

\item{restrictToDomain}{Flag that determines whether responses are sent to recipients who are outside of the user's domain}

\item{startTime}{An optional start time for sending auto-replies (epoch ms)}
}
\value{
VacationSettings object
}
\description{
VacationSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Vacation auto-reply settings for an account. These settings correspond to the 'Vacation responder' feature in the web interface.
}
