% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{WatchRequest}
\alias{WatchRequest}
\title{WatchRequest Object}
\usage{
WatchRequest(labelFilterAction = NULL, labelIds = NULL, topicName = NULL)
}
\arguments{
\item{labelFilterAction}{Filtering behavior of labelIds list specified}

\item{labelIds}{List of label_ids to restrict notifications about}

\item{topicName}{A fully qualified Google Cloud Pub/Sub API topic name to publish the events to}
}
\value{
WatchRequest object
}
\description{
WatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Set up or update a new push notification watch on this user's mailbox.
}
\seealso{
Other WatchRequest functions: \code{\link{users.watch}}
}
