% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_objects.R
\name{WatchResponse}
\alias{WatchResponse}
\title{WatchResponse Object}
\usage{
WatchResponse(expiration = NULL, historyId = NULL)
}
\arguments{
\item{expiration}{When Gmail will stop sending notifications for mailbox updates (epoch millis)}

\item{historyId}{The ID of the mailbox's current history record}
}
\value{
WatchResponse object
}
\description{
WatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Push notification watch response.
}
